Pass times as double, not as integer

+++ util.h
@@ -15,7 +15,7 @@ void handle_srt_for_file (struct atrfs_entry *file, bool insert);
 
 void tmplog(char *fmt, ...);
 void get_all_file_entries (struct atrfs_entry ***entries, size_t *count);
-char *get_pdir(int filelen, int watchcount, int watchtime);
+char *get_pdir(double filelen, int watchcount, double watchtime);
 char *secs_to_time (double secs);
 char *pid_to_cmdline(pid_t pid);
 double doubletime(void);
